restructure paginator
